created Dtos for entities
