+++ dotnetflix.Models/Dtos/Show/ShowDto.cs
@@ -1,14 +1,14 @@
-namespace dotnetflix.Models.Dtos;
+namespace dotnetflix.Models.Dtos.Show;
 
 public class ShowDto
 {
     public int Id { get; set; }
     public int MovieId { get; set; }
-    // public string Movie { get; set; }
+    public string MovieTitle { get; set; }
     public int TheaterId { get; set; }
-    // public string Theater { get; set; }
+    public string TheaterName { get; set; }
     public DateTime Date { get; set; }
-    public TimeSpan Time { get; set; }
+    public int ScreenTime { get; set; }
     public string Type { get; set; }
     public decimal BasePrice { get; set; }
 }
